+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1
 * Added POT file for translations
+* Option for enhanced link attribution
 
 ## 0.1
 * First version